Restructure RuneliteObject usage
Merges Cheerer with the overall RuneliteObjectMananger.
