+++ src/main/java/com/questhelper/helpers/quests/deserttreasureii/FakeWhisperer.java
@@ -30,13 +30,13 @@ import com.questhelper.requirements.conditional.ObjectCondition;
 import static com.questhelper.requirements.util.LogicHelper.nor;
 import com.questhelper.requirements.util.Operation;
 import com.questhelper.requirements.var.VarbitRequirement;
-import com.questhelper.steps.playermadesteps.RuneliteDialogStep;
-import com.questhelper.steps.playermadesteps.RuneliteObjectDialogStep;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.FaceAnimationIDs;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.FakeGraphicsObject;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.FakeNpc;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.RuneliteObjectManager;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.actions.Action;
+import com.questhelper.runeliteobjects.dialog.RuneliteDialogStep;
+import com.questhelper.runeliteobjects.dialog.RuneliteObjectDialogStep;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.FaceAnimationIDs;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.FakeGraphicsObject;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.FakeNpc;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.RuneliteObjectManager;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.actions.Action;
 import com.questhelper.steps.tools.QuestPerspective;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
