+++ src/main/java/com/questhelper/playerquests/cookshelper/CooksHelper.java
@@ -40,15 +40,15 @@ import com.questhelper.steps.ConditionalStep;
 import com.questhelper.steps.DetailedQuestStep;
 import com.questhelper.steps.QuestStep;
 import com.questhelper.steps.TileStep;
-import com.questhelper.steps.playermadesteps.RuneliteConfigSetter;
-import com.questhelper.steps.playermadesteps.RuneliteDialogStep;
-import com.questhelper.steps.playermadesteps.RuneliteObjectDialogStep;
+import com.questhelper.runeliteobjects.RuneliteConfigSetter;
+import com.questhelper.runeliteobjects.dialog.RuneliteDialogStep;
+import com.questhelper.runeliteobjects.dialog.RuneliteObjectDialogStep;
 import com.questhelper.steps.playermadesteps.RuneliteObjectStep;
-import com.questhelper.steps.playermadesteps.RunelitePlayerDialogStep;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.FaceAnimationIDs;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.FakeItem;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.FakeNpc;
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.ReplacedObject;
+import com.questhelper.runeliteobjects.dialog.RunelitePlayerDialogStep;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.FaceAnimationIDs;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.FakeItem;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.FakeNpc;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.ReplacedObject;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
