+++ src/main/java/com/questhelper/questhelpers/PlayerMadeQuestHelper.java
@@ -24,7 +24,7 @@
  */
 package com.questhelper.questhelpers;
 
-import com.questhelper.steps.playermadesteps.extendedruneliteobjects.QuestCompletedWidget;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.QuestCompletedWidget;
 import javax.inject.Inject;
 import lombok.Getter;
 import net.runelite.api.QuestState;
