+++ src/main/java/com/questhelper/runeliteobjects/Cheerer.java
@@ -25,124 +25,111 @@
  */
 package com.questhelper.runeliteobjects;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Random;
-import java.util.function.Function;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import net.runelite.api.Animation;
-import net.runelite.api.ChatMessageType;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.FakeNpc;
+import com.questhelper.runeliteobjects.extendedruneliteobjects.RuneliteObjectManager;
 import net.runelite.api.Client;
 import net.runelite.api.JagexColor;
 import net.runelite.api.Model;
 import net.runelite.api.ModelData;
-import net.runelite.api.RuneLiteObject;
-import net.runelite.api.coords.LocalPoint;
 import net.runelite.api.coords.WorldPoint;
-import java.awt.Color;
-import net.runelite.client.callback.ClientThread;
-import net.runelite.client.chat.ChatColorType;
-import net.runelite.client.chat.ChatMessageBuilder;
+import java.awt.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
 import net.runelite.client.chat.ChatMessageManager;
-import net.runelite.client.chat.QueuedMessage;
+import net.runelite.client.config.ConfigManager;
 
 public class Cheerer
 {
-	public final RuneLiteObject runeLiteObject;
+	private static final List<FakeNpc> cheerers = new ArrayList<>();
 
-	public final WorldPoint worldPoint;
-	private final Client client;
-	private final ClientThread clientThread;
-	@Getter
-	private final Style style;
-
-	@Getter
-	private final String message;
-	private static final int QUEST_HOOD_MALE = 18914;
-	private static final int QUEST_CAPE_MALE = 18946;
-
-	@RequiredArgsConstructor
-	public enum Style
+	public static void createCheerers(RuneliteObjectManager runeliteObjectManager, Client client, ConfigManager configManager)
 	{
-		FAN(l ->
-		{
-			short qpcDark = (short) -22440;
-
-			ModelData hood = l.client.loadModelData(QUEST_HOOD_MALE).cloneColors()
-				.recolor((short) -21568, qpcDark) // Inside
-				.recolor((short) 22464, qpcDark)
-				.recolor((short) 960, qpcDark);
-
-			ModelData cape = l.client.loadModelData(QUEST_CAPE_MALE).cloneColors()
-				.recolor((short) -8256, qpcDark) // Outside trim
-				.recolor((short) -11353, JagexColor.rgbToHSL(Color.WHITE.getRGB(), 1.0d)); // Inside trim
-
-			ModelData mdf = createModel(l.client,
-				11359, 38101, 38079, 4925, 10706, 358);
-			mdf = createModel(l.client, mdf, hood, cape);
-
-			return mdf.cloneColors()
-				.light(10 + ModelData.DEFAULT_AMBIENT, 1875 + ModelData.DEFAULT_CONTRAST,
-					ModelData.DEFAULT_X, ModelData.DEFAULT_Y, 20);
-		}, anim(862), "Zoinkwiz", "Loves questing."),
-		WISE_OLD_MAN(l ->
-		{
-			short clothingColor = JagexColor.rgbToHSL(new Color(102, 93, 44).getRGB(), 0.6d);
-			short blue = JagexColor.rgbToHSL(Color.BLUE.getRGB(), 1.0d);
-			ModelData skirt = l.client.loadModelData(265).cloneColors()
-				.recolor((short)25238, clothingColor);
-
-			ModelData shirt = l.client.loadModelData(292).cloneColors()
-				.recolor((short)8741, clothingColor);
-
-			ModelData arms = l.client.loadModelData(170).cloneColors()
-				.recolor((short)8741, clothingColor);
-
-			// 8741
-			ModelData cape = l.client.loadModelData(323).cloneColors()
-				.recolor((short) 926, blue) // Inside
-				.recolor((short) 7700, blue) // Mail cape
-				.recolor((short) 11200, (short)8741); // Trim
-
-			ModelData partyhat = l.client.loadModelData(187).cloneColors()
-				.recolor((short)926, blue);
-			ModelData mdf = createModel(l.client,
-				9103, // face
-				4925, // beard
-				176, // hands
-				181); // feet?
-
-			mdf = createModel(l.client, mdf, skirt, shirt, arms, cape, partyhat);
-
-			return mdf.cloneColors()
-				.light();
-		}, anim(862), "Wise Old Man", "Loves questing."),
-		;
-
-		private final Function<Cheerer, Model> modelSupplier;
-		private final Function<Cheerer, Animation> animationSupplier;
-
-		@Getter
-		private final String displayName;
-
-		@Getter
-		private final String examine;
-
-		private static final List<Style> VALUES = Collections.unmodifiableList(Arrays.asList(values()));
-		private static final int SIZE = VALUES.size();
-		private static final Random RANDOM = new Random();
-
-		public static Style randomCheerer()
-		{
-			return VALUES.get(RANDOM.nextInt(SIZE));
-		}
+		createWOM(runeliteObjectManager, client);
+		createZoinkwiz(runeliteObjectManager, client);
 	}
 
-	private static Function<Cheerer, Animation> anim(int id)
+	private static void createWOM(RuneliteObjectManager runeliteObjectManager, Client client)
 	{
-		return b -> b.client.loadAnimation(id);
+		WorldPoint playerPos = client.getLocalPlayer().getWorldLocation();
+		WorldPoint pointAbovePlayer = new WorldPoint(playerPos.getX(), playerPos.getY() + 1, playerPos.getPlane());
+		FakeNpc wiseOldMan = runeliteObjectManager.createFakeNpc("global", wiseOldManOutfit(client), pointAbovePlayer, 862);
+		wiseOldMan.setName("Wise Old Man");
+		wiseOldMan.setExamine("Loves questing.");
+		wiseOldMan.addExamineAction(runeliteObjectManager);
+		wiseOldMan.disable();
+
+		cheerers.add(wiseOldMan);
+	}
+
+	private static Model wiseOldManOutfit(Client client)
+	{
+		short clothingColor = JagexColor.rgbToHSL(new Color(102, 93, 44).getRGB(), 0.6d);
+		short blue = JagexColor.rgbToHSL(Color.BLUE.getRGB(), 1.0d);
+		ModelData skirt = client.loadModelData(265).cloneColors()
+			.recolor((short)25238, clothingColor);
+
+		ModelData shirt = client.loadModelData(292).cloneColors()
+			.recolor((short)8741, clothingColor);
+
+		ModelData arms = client.loadModelData(170).cloneColors()
+			.recolor((short)8741, clothingColor);
+
+		// 8741
+		ModelData cape = client.loadModelData(323).cloneColors()
+			.recolor((short) 926, blue) // Inside
+			.recolor((short) 7700, blue) // Mail cape
+			.recolor((short) 11200, (short) 8741); // Trim
+
+		ModelData partyhat = client.loadModelData(187).cloneColors()
+			.recolor((short)926, blue);
+		ModelData mdf = createModel(client,
+			9103, // face
+			4925, // beard
+			176, // hands
+			181); // feet?
+
+		mdf = createModel(client, mdf, skirt, shirt, arms, cape, partyhat);
+
+		return mdf.cloneColors()
+			.light();
+	}
+
+	private static void createZoinkwiz(RuneliteObjectManager runeliteObjectManager, Client client)
+	{
+		WorldPoint playerPos = client.getLocalPlayer().getWorldLocation();
+		WorldPoint pointAbovePlayer = new WorldPoint(playerPos.getX(), playerPos.getY() + 1, playerPos.getPlane());
+		FakeNpc zoinkwiz = runeliteObjectManager.createFakeNpc("global", zoinkwizOutfit(client), pointAbovePlayer, 862);
+		zoinkwiz.setName("Zoinkwiz");
+		zoinkwiz.setExamine("Loves questing.");
+		zoinkwiz.addExamineAction(runeliteObjectManager);
+		zoinkwiz.disable();
+
+		cheerers.add(zoinkwiz);
+	}
+
+	private static Model zoinkwizOutfit(Client client)
+	{
+		short qpcDark = (short) -22440;
+		final int QUEST_HOOD_MALE = 18914;
+		final int QUEST_CAPE_MALE = 18946;
+
+		ModelData hood = client.loadModelData(QUEST_HOOD_MALE).cloneColors()
+			.recolor((short) -21568, qpcDark) // Inside
+			.recolor((short) 22464, qpcDark)
+			.recolor((short) 960, qpcDark);
+
+		ModelData cape = client.loadModelData(QUEST_CAPE_MALE).cloneColors()
+			.recolor((short) -8256, qpcDark) // Outside trim
+			.recolor((short) -11353, JagexColor.rgbToHSL(Color.WHITE.getRGB(), 1.0d)); // Inside trim
+
+		ModelData mdf = createModel(client,
+			11359, 38101, 38079, 4925, 10706, 358);
+		mdf = createModel(client, mdf, hood, cape);
+
+		return mdf.cloneColors()
+			.light(10 + ModelData.DEFAULT_AMBIENT, 1875 + ModelData.DEFAULT_CONTRAST,
+				ModelData.DEFAULT_X, ModelData.DEFAULT_Y, 20);
 	}
 
 	private static ModelData createModel(Client client, ModelData... data)
@@ -160,58 +147,18 @@ public class Cheerer
 		return client.mergeModels(modelData);
 	}
 
-	public Cheerer(Client client, ClientThread clientThread, WorldPoint worldPoint, Style style, ChatMessageManager chatMessageManager, String message)
+	public static void activateCheerer(Client client, ChatMessageManager chatMessageManager)
 	{
-		this.client = client;
-		this.clientThread = clientThread;
-		this.style = style;
-		this.runeLiteObject = client.createRuneLiteObject();
-		this.worldPoint = worldPoint;
-		this.message = message;
+		Random generator = new Random();
+		FakeNpc cheerer = cheerers.get(generator.nextInt(cheerers.size()));
+		WorldPoint playerPos = client.getLocalPlayer().getWorldLocation();
 
-		update();
+		int tickToRemoveOn = client.getTickCount() + 200;
 
-		runeLiteObject.setShouldLoop(true);
-
-		LocalPoint lp = LocalPoint.fromWorld(client, this.worldPoint);
-
-		String chatMessage = new ChatMessageBuilder()
-			.append(ChatColorType.NORMAL)
-			.append(this.message)
-			.build();
-
-		chatMessageManager.queue(QueuedMessage.builder()
-			.type(ChatMessageType.PUBLICCHAT)
-			.name(style.getDisplayName())
-			.runeLiteFormattedMessage(chatMessage)
-			.timestamp((int) (System.currentTimeMillis() / 1000))
-			.build());
-		runeLiteObject.setLocation(lp, client.getPlane());
-
-		runeLiteObject.setActive(true);
+		WorldPoint pointAbovePlayer = new WorldPoint(playerPos.getX(), playerPos.getY() + 1, playerPos.getPlane());
+		cheerer.setWorldPoint(pointAbovePlayer);
+		cheerer.activate();
+		cheerer.addOverheadText("Congratz on completing the quest!", tickToRemoveOn, chatMessageManager);
+		cheerer.setDisableAfterTick(tickToRemoveOn);
 	}
-
-	private void update()
-	{
-		clientThread.invoke(() ->
-		{
-			Model model = style.modelSupplier.apply(this);
-			if (model == null)
-			{
-				return false;
-			}
-
-			Animation anim = style.animationSupplier.apply(this);
-
-			runeLiteObject.setAnimation(anim);
-			runeLiteObject.setModel(model);
-			return true;
-		});
-	}
-
-	public void remove()
-	{
-		runeLiteObject.setActive(false);
-	}
-
 }
