+++ src/main/java/com/questhelper/runeliteobjects/dialog/RuneliteObjectDialogStep.java
@@ -22,85 +22,42 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.questhelper.steps.playermadesteps;
+package com.questhelper.runeliteobjects.dialog;
 
+import com.questhelper.runeliteobjects.extendedruneliteobjects.FaceAnimationIDs;
 import java.util.ArrayList;
+import com.questhelper.runeliteobjects.RuneliteConfigSetter;
 import lombok.Getter;
-import lombok.Setter;
+import net.runelite.api.Client;
 
-public class RuneliteDialogStep
+public class RuneliteObjectDialogStep extends RuneliteDialogStep
 {
-	@Getter
-	protected final String name;
-
-	@Getter
-	protected final String text;
-
-	// Only used for NPCs
-	@Getter
-	protected final int faceID;
-
-	@Getter
-	protected final int animation;
-
-	@Getter
-	private RuneliteDialogStep continueDialog;
-
-	@Getter
-	private RuneliteConfigSetter stateProgression;
-
 	@Getter
 	private final ArrayList<RuneliteDialogStep> dialogChoices = new ArrayList<>();
 
-	public RuneliteDialogStep(String name, String text, int faceID, int animation)
+	public RuneliteObjectDialogStep(Client client, String text, FaceAnimationIDs animation)
 	{
-		this.name = name;
-		this.text = text;
-		this.animation = animation;
-		this.faceID = faceID;
+		super(client.getLocalPlayer().getName(), text, -1, animation.getAnimationID());
+		client.getLocalPlayer().getName();
 	}
 
-	public RuneliteDialogStep(String name, String text, int faceID, int animation, RuneliteConfigSetter setter)
+	public RuneliteObjectDialogStep(String name, String text, int faceID, int animation)
 	{
-		this.name = name;
-		this.text = text;
-		this.animation = animation;
-		this.faceID = faceID;
-		this.setStateProgression(setter);
+		super(name, text, faceID, animation);
 	}
 
-	public RuneliteDialogStep setStateProgression(RuneliteConfigSetter stateProgression)
+	public RuneliteObjectDialogStep(String name, String text, int faceID, int animation, RuneliteConfigSetter setter)
 	{
-		this.stateProgression = stateProgression;
-		return this;
+		super(name, text, faceID, animation, setter);
 	}
 
-	public void addNewDialogChoice(RuneliteDialogStep step)
+	public RuneliteObjectDialogStep(String name, String text, int faceID, RuneliteConfigSetter setter)
 	{
-		dialogChoices.add(step);
+		super(name, text, faceID, 570, setter);
 	}
 
-	public boolean isPlayer()
+	public RuneliteObjectDialogStep(String name, String text, int faceID)
 	{
-		return false;
-	}
-
-	public boolean isStateChanger()
-	{
-		return stateProgression != null;
-	}
-
-	public void progressState()
-	{
-		if (isStateChanger())
-		{
-			stateProgression.setConfigValue();
-		}
-	}
-
-	public RuneliteDialogStep addContinueDialog(RuneliteDialogStep continueDialog)
-	{
-		this.continueDialog = continueDialog;
-		return continueDialog;
+		this(name, text, faceID, 570);
 	}
 }
