+++ src/main/java/com/questhelper/runeliteobjects/extendedruneliteobjects/ChatBox.java
@@ -23,9 +23,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.questhelper.steps.playermadesteps.extendedruneliteobjects;
+package com.questhelper.runeliteobjects.extendedruneliteobjects;
 
-import com.questhelper.steps.playermadesteps.RuneliteDialogStep;
+import com.questhelper.runeliteobjects.dialog.RuneliteDialogStep;
 import java.awt.event.KeyEvent;
 import java.util.ArrayList;
 import java.util.List;
