+++ src/main/java/com/questhelper/runeliteobjects/extendedruneliteobjects/FakeGraphicsObject.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2023, Zoinkwiz <https://github.com/Zoinkwiz>
- * Copyright (c) 2018 Abex
+ * Copyright (c) 2023, Zoinkwiz
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -23,15 +22,45 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.questhelper.steps.playermadesteps.extendedruneliteobjects;
+package com.questhelper.runeliteobjects.extendedruneliteobjects;
 
 import net.runelite.api.Client;
-import net.runelite.client.game.chatbox.ChatboxPanelManager;
+import net.runelite.api.coords.WorldPoint;
+import net.runelite.client.callback.ClientThread;
 
-public class NpcChatBox extends ChatBox
+public class FakeGraphicsObject extends ExtendedRuneliteObject
 {
-	protected NpcChatBox(Client client, ChatboxPanelManager chatboxPanelManager)
+	ExtendedRuneliteObject objectToSpawnAfter;
+	protected FakeGraphicsObject(Client client, ClientThread clientThread, WorldPoint worldPoint,
+								 int[] model, int animation, ExtendedRuneliteObject objectToSpawnAfter)
 	{
-		super(client, chatboxPanelManager);
+		super(client, clientThread, worldPoint, model, animation);
+		objectType = RuneliteObjectTypes.GRAPHICS_OBJECT;
+		this.objectToSpawnAfter = objectToSpawnAfter;
+		runeliteObject.setShouldLoop(false);
+		runeliteObject.setActive(false);
+	}
+
+	protected FakeGraphicsObject(Client client, ClientThread clientThread, WorldPoint worldPoint,
+								 int[] model, int animation)
+	{
+		super(client, clientThread, worldPoint, model, animation);
+		objectType = RuneliteObjectTypes.GRAPHICS_OBJECT;
+		this.objectToSpawnAfter = null;
+		runeliteObject.setShouldLoop(false);
+		runeliteObject.setActive(false);
+	}
+
+	@Override
+	protected void actionOnClientTick()
+	{
+		if (runeliteObject.getAnimation().getNumFrames() <= runeliteObject.getAnimationFrame() + 1)
+		{
+			disable();
+			if (objectToSpawnAfter != null)
+			{
+				objectToSpawnAfter.activate();
+			}
+		}
 	}
 }
