+++ src/main/java/com/questhelper/runeliteobjects/extendedruneliteobjects/PlayerChatBox.java
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2023, Zoinkwiz <https://github.com/Zoinkwiz>
+ * Copyright (c) 2018 Abex
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,18 +23,45 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.questhelper.steps.playermadesteps.extendedruneliteobjects;
+package com.questhelper.runeliteobjects.extendedruneliteobjects;
+
 
 import net.runelite.api.Client;
-import net.runelite.api.coords.WorldPoint;
-import net.runelite.client.callback.ClientThread;
+import net.runelite.api.widgets.Widget;
+import net.runelite.api.widgets.WidgetModelType;
+import net.runelite.api.widgets.WidgetPositionMode;
+import net.runelite.api.widgets.WidgetType;
+import net.runelite.client.game.chatbox.ChatboxPanelManager;
 
-public class FakeObject extends ExtendedRuneliteObject
+public class PlayerChatBox extends ChatBox
 {
-	protected FakeObject(Client client, ClientThread clientThread, WorldPoint worldPoint, int[] model, int animation)
+	protected PlayerChatBox(Client client, ChatboxPanelManager chatboxPanelManager)
 	{
-		super(client, clientThread, worldPoint, model, animation);
-		objectType = RuneliteObjectTypes.OBJECT;
-		nameColor = "00FFFF";
+		super(client, chatboxPanelManager);
+	}
+
+	@Override
+	protected void open()
+	{
+		Widget container = chatboxPanelManager.getContainerWidget();
+
+		Widget npcFaceWidget = container.createChild(0, WidgetType.RECTANGLE);
+		npcFaceWidget.setType(WidgetType.MODEL);
+		npcFaceWidget.setModelId(dialog.getFaceID());
+		npcFaceWidget.setAnimationId(dialog.getAnimation());
+		npcFaceWidget.setRotationX(40);
+		npcFaceWidget.setRotationZ(166);
+		npcFaceWidget.setModelZoom(796);
+		npcFaceWidget.setOriginalX(420);
+		npcFaceWidget.setOriginalY(50);
+		npcFaceWidget.setOriginalWidth(32);
+		npcFaceWidget.setOriginalHeight(32);
+		npcFaceWidget.setXPositionMode(WidgetPositionMode.ABSOLUTE_TOP);
+		npcFaceWidget.setYPositionMode(WidgetPositionMode.ABSOLUTE_TOP);
+		npcFaceWidget.setModelType(WidgetModelType.LOCAL_PLAYER_CHATHEAD);
+		npcFaceWidget.revalidate();
+
+		setupDialog(container, 5);
 	}
 }
+
